Update counter-with-dancer-using-in-memory-sqlite-database.txt
Referred here by https://perlmaven.com/counter-with-dancer-using-in-memory-sqlite-database

The code is unclear, especially as it's not explained.

(1) Is the .psgi extension necessary?  Can .cgi or .pl be used?
(2) On line 5, is the database created the first time the script is run, and then the script just connects to the database on subsequent runs?
(3) On line 18, where is $counter returned to?
(4) There is no print command.  How is there any output?

Code raises more questions than it answers.